requirements.txt
django>1.10
djangorestframework
djangorestframework-gis
django-rest-swagger
django-bootstrap3
django-sekizai
django-webpack-loader
channels
asgi_redis
psycopg2
gunicorn
